last features

+++ ITHelpDesk.Application/Services/CommentService.cs
@@ -1,8 +1,10 @@
-﻿using ITHelpDesk.Application.DTOs.TicketComments;
+using FluentValidation;
+using ITHelpDesk.Application.DTOs.TicketComments;
 using ITHelpDesk.Application.Interfaces.Identity;
 using ITHelpDesk.Application.Interfaces.Services;
 using ITHelpDesk.Application.Interfaces.UnitOfWork;
 using ITHelpDesk.Domain.Entities;
+using ITHelpDesk.Domain.Enums;
 using System.Net.Sockets;
 
 namespace ITHelpDesk.Application.Services
@@ -10,20 +12,35 @@ namespace ITHelpDesk.Application.Services
     public class CommentService : ICommentService
     {
         private readonly IUnitOfWork _unitOfWork;
+        private readonly IHistoryService _historyService;
         private readonly IUserService _userService;
         private readonly ICurrentUserService _currentUserService;
+        private readonly IValidator<CreateTicketCommentDto> _createValidator;
+        private readonly IValidator<EditTicketCommentDto> _editValidator;
 
-        public CommentService(IUnitOfWork unitOfWork, ICurrentUserService currentUserService, IUserService userService)
+        public CommentService(IUnitOfWork unitOfWork, ICurrentUserService currentUserService,
+                              IUserService userService, IHistoryService historyService,
+                              IValidator<CreateTicketCommentDto> createValidator,
+                              IValidator<EditTicketCommentDto> editValidator)
         {
             _unitOfWork = unitOfWork;
             _currentUserService = currentUserService;
+            _historyService = historyService;
             _userService = userService;
+            _createValidator = createValidator;
+            _editValidator = editValidator;
         }
 
         public async Task<GeneralResult> CreateCommentAsync(CreateTicketCommentDto dto)
         {
-            var ticket =
-                await _unitOfWork.TicketRepository.GetByIdAsync(dto.TicketId);
+            var validationResult = await _createValidator.ValidateAsync(dto);
+            if (!validationResult.IsValid)
+            {
+                Dictionary<string, List<Error>> errors = validationResult.ToError();
+                return GeneralResult.FailedResult(errors);
+            }
+
+            var ticket = await _unitOfWork.TicketRepository.GetByIdAsync(dto.TicketId);
 
             if (ticket == null)
                 return GeneralResult.NotFound("Ticket not found.");
@@ -40,57 +57,82 @@ namespace ITHelpDesk.Application.Services
             };
 
             await _unitOfWork.TicketCommentRepository.AddAsync(comment);
+            await _historyService.LogHistoryAsync(ticket, TicketHistoryField.CommentAdded);
+
             await _unitOfWork.SaveChangesAsync();
 
             return GeneralResult.SuccessedResult("Comment added.");
         }
+
         public async Task<GeneralResult<int>> EditCommentAsync(EditTicketCommentDto dto)
         {
-            var comment = await _unitOfWork.TicketCommentRepository.GetByIdAsync(dto.CommentId);
+            var validationResult = await _editValidator.ValidateAsync(dto);
+            if (!validationResult.IsValid)
+            {
+                Dictionary<string, List<Error>> errors = validationResult.ToError();
+                return GeneralResult<int>.FailedResult(errors);
+            }
 
+            var comment = await _unitOfWork.TicketCommentRepository.GetCommentWithTicketByIdAsync(dto.CommentId);
             if (comment == null)
                 return GeneralResult<int>.NotFound("Comment not found.");
 
+            if (comment.CreatedById != _currentUserService.UserId)
+            {
+                return GeneralResult<int>.FailedResult("You do not have permission to edit this ticket.");
+            }
+
             if (comment.Content == dto.Content)
                 return GeneralResult<int>.SuccessedResult(comment.TicketId, "No changes detected.");
+
+            await _historyService.LogHistoryAsync(comment.Ticket!, TicketHistoryField.CommentEdited, comment.Content, dto.Content);
 
             comment.Content = dto.Content;
             comment.UpdatedAt = DateTime.UtcNow;
 
             _unitOfWork.TicketCommentRepository.Update(comment);
-
             await _unitOfWork.SaveChangesAsync();
 
             return GeneralResult<int>.SuccessedResult(comment.TicketId, "Comment updated.");
         }
+
         public async Task<GeneralResult> DeleteCommentAsync(int commentId)
         {
-            var comment =
-                await _unitOfWork.TicketCommentRepository.GetByIdAsync(commentId);
-
+            var comment = await _unitOfWork.TicketCommentRepository.GetCommentWithTicketByIdAsync(commentId);
             if (comment == null)
                 return GeneralResult.NotFound("Comment not found.");
 
+            if (!_currentUserService.IsAdmin && comment.CreatedById != _currentUserService.UserId)
+            {
+                return GeneralResult.FailedResult("You do not have permission to edit this ticket.");
+            }
+
             _unitOfWork.TicketCommentRepository.Remove(comment);
 
+            await _historyService.LogHistoryAsync(comment.Ticket!, TicketHistoryField.CommentDeleted);
             await _unitOfWork.SaveChangesAsync();
+
 
             return GeneralResult.SuccessedResult("Comment deleted.");
         }
-        public async Task<IEnumerable<TicketCommentDto>> GetTicketCommentsAsync(int ticketId)
+
+        public async Task<List<TicketCommentDto>> GetTicketCommentsAsync(int ticketId)
         {
             var ticketComments = await _unitOfWork.TicketCommentRepository.GetCommentsForTicketAsync(ticketId);
+            
+            var userIds = ticketComments.Select(c => c.CreatedById).Distinct().ToList();
+            var userNames = await _userService.GetUserNamesByIdsAsync(userIds);
 
-            var result = await Task.WhenAll(
-                ticketComments.Select(async c => new TicketCommentDto(
+            var result = ticketComments.Select(c => new TicketCommentDto(
                     c.CommentId,
-                    await _userService.GetUserNameByIdAsync(c.CreatedById) ?? "Unknown",
+                    c.CreatedById,
+                    userNames.GetValueOrDefault(c.CreatedById, "Unknown"),
                     c.Content,
                     c.CreatedAt,
                     c.UpdatedAt
-                )));
+                ));
 
-            return result;
+            return result.ToList();
         }
     }
 }
